+++ checker-qual/src/main/java/org/checkerframework/checker/calledmethods/qual/EnsuresCalledMethods.java
@@ -40,11 +40,13 @@ import java.lang.annotation.Target;
  * </pre>
  *
  * @see EnsuresCalledMethodsIf
+ * @see EnsuresCalledMethodsOnException
  * @checker_framework.manual #called-methods-checker Called Methods Checker
  */
 @PostconditionAnnotation(qualifier = CalledMethods.class)
 @Target({ElementType.METHOD, ElementType.CONSTRUCTOR})
 @Repeatable(EnsuresCalledMethods.List.class)
+@InheritedAnnotation
 public @interface EnsuresCalledMethods {
     /**
      * The Java expressions that will have methods called on them.
@@ -73,6 +75,7 @@ public @interface EnsuresCalledMethods {
     @Retention(RetentionPolicy.RUNTIME)
     @Target({ElementType.METHOD, ElementType.CONSTRUCTOR})
     @InheritedAnnotation
+    @PostconditionAnnotation(qualifier = CalledMethods.class)
     public static @interface List {
         /**
          * Return the repeatable annotations.