+++ checker/src/main/java/org/checkerframework/checker/calledmethods/CalledMethodsTransfer.java
@@ -15,8 +15,11 @@ import org.checkerframework.dataflow.cfg.node.Node;
 import org.checkerframework.dataflow.expression.JavaExpression;
 import org.checkerframework.framework.flow.CFAbstractStore;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
+import org.checkerframework.framework.util.JavaExpressionParseUtil;
+import org.checkerframework.framework.util.StringToJavaExpression;
 import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.AnnotationUtils;
+import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.TreeUtils;
 import org.plumelib.util.CollectionsPlume;
 
@@ -26,10 +29,12 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeMirror;
+import javax.lang.model.util.Types;
 
 /** A transfer function that accumulates the names of methods called. */
 public class CalledMethodsTransfer extends AccumulationTransfer {
@@ -52,6 +57,16 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
      */
     private final ExecutableElement calledMethodsValueElement;
 
+    /** The type mirror for {@link Exception}. */
+    protected final TypeMirror javaLangExceptionType;
+
+    /* NO-AFU
+     * True if -AenableWpiForRlc was passed on the command line. See {@link
+     * ResourceLeakChecker#ENABLE_WPI_FOR_RLC}.
+     *
+    private final boolean enableWpiForRlc;
+    */
+
     /**
      * Create a new CalledMethodsTransfer.
      *
@@ -61,7 +76,51 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
         super(analysis);
         calledMethodsValueElement =
                 ((CalledMethodsAnnotatedTypeFactory) atypeFactory).calledMethodsValueElement;
+        // enableWpiForRlc =
+        // atypeFactory.getChecker().hasOption(ResourceLeakChecker.ENABLE_WPI_FOR_RLC);
+
+        ProcessingEnvironment env = atypeFactory.getProcessingEnv();
+        javaLangExceptionType =
+                env.getTypeUtils()
+                        .getDeclaredType(ElementUtils.getTypeElement(env, Exception.class));
     }
+
+    /* NO-AFU
+     * @param tree a tree
+     * @return false if Resource Leak Checker is running as one of the upstream checkers and the
+     *     -AenableWpiForRlc flag (see {@link ResourceLeakChecker#ENABLE_WPI_FOR_RLC}) is not passed
+     *     as a command line argument, otherwise returns the result of the super call
+     */
+    /* NO-AFU
+    @Override
+    protected boolean shouldPerformWholeProgramInference(Tree tree) {
+      if (!isWpiEnabledForRLC()
+          && atypeFactory.getCheckerNames().contains(ResourceLeakChecker.class.getCanonicalName())) {
+        return false;
+      }
+      return super.shouldPerformWholeProgramInference(tree);
+    }
+    */
+
+    /* NO-AFU
+     * See {@link ResourceLeakChecker#ENABLE_WPI_FOR_RLC}.
+     *
+     * @param expressionTree a tree
+     * @param lhsTree its element
+     * @return false if Resource Leak Checker is running as one of the upstream checkers and the
+     *     -AenableWpiForRlc flag is not passed as a command line argument, otherwise returns the
+     *     result of the super call
+     */
+    /* NO-AFU
+    @Override
+    protected boolean shouldPerformWholeProgramInference(Tree expressionTree, Tree lhsTree) {
+      if (!isWpiEnabledForRLC()
+          && atypeFactory.getCheckerNames().contains(ResourceLeakChecker.class.getCanonicalName())) {
+        return false;
+      }
+      return super.shouldPerformWholeProgramInference(expressionTree, lhsTree);
+    }
+    */
 
     @Override
     public TransferResult<AccumulationValue, AccumulationStore> visitMethodInvocation(
@@ -69,7 +128,11 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
         exceptionalStores = makeExceptionalStores(node, input);
         TransferResult<AccumulationValue, AccumulationStore> superResult =
                 super.visitMethodInvocation(node, input);
-        handleEnsuresCalledMethodsVarArgs(node, superResult);
+
+        ExecutableElement method = TreeUtils.elementFromUse(node.getTree());
+        handleEnsuresCalledMethodsVarArgs(node, method, superResult);
+        handleEnsuresCalledMethodsOnException(node, method, exceptionalStores);
+
         Node receiver = node.getTarget().getReceiver();
         if (receiver != null) {
             String methodName = node.getTarget().getMethod().getSimpleName().toString();
@@ -120,7 +183,12 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
                 }
             }
             AnnotationMirror newAnno = atypeFactory.createAccumulatorAnnotation(valuesAsList);
-            exceptionalStores.forEach((tm, s) -> s.insertValue(target, newAnno));
+            exceptionalStores.forEach(
+                    (tm, s) ->
+                            s.replaceValue(
+                                    target,
+                                    analysis.createSingleAnnotationValue(
+                                            newAnno, target.getType())));
         }
     }
 
@@ -155,12 +223,13 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
      * present.
      *
      * @param node the method invocation node
+     * @param elt the method being invoked
      * @param result the current result
      */
     private void handleEnsuresCalledMethodsVarArgs(
             MethodInvocationNode node,
+            ExecutableElement elt,
             TransferResult<AccumulationValue, AccumulationStore> result) {
-        ExecutableElement elt = TreeUtils.elementFromUse(node.getTree());
         AnnotationMirror annot =
                 atypeFactory.getDeclAnnotation(elt, EnsuresCalledMethodsVarArgs.class);
         if (annot == null) {
@@ -193,6 +262,53 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
                 JavaExpression receiverReceiver = JavaExpression.fromNode(arg);
                 thenStore.insertValue(receiverReceiver, newType);
                 elseStore.insertValue(receiverReceiver, newType);
+            }
+        }
+    }
+
+    /**
+     * Update the given <code>exceptionalStores</code> for the {@link
+     * org.checkerframework.checker.calledmethods.qual.EnsuresCalledMethodsOnException} annotations
+     * written on the given <code>method</code>.
+     *
+     * @param node a method invocation
+     * @param method the method being invoked
+     * @param exceptionalStores the stores to update
+     */
+    private void handleEnsuresCalledMethodsOnException(
+            MethodInvocationNode node,
+            ExecutableElement method,
+            Map<TypeMirror, AccumulationStore> exceptionalStores) {
+        Types types = atypeFactory.getProcessingEnv().getTypeUtils();
+        for (EnsuresCalledMethodOnExceptionContract postcond :
+                ((CalledMethodsAnnotatedTypeFactory) atypeFactory)
+                        .getExceptionalPostconditions(method)) {
+            JavaExpression e;
+            try {
+                e =
+                        StringToJavaExpression.atMethodInvocation(
+                                postcond.getExpression(),
+                                node.getTree(),
+                                atypeFactory.getChecker());
+            } catch (JavaExpressionParseUtil.JavaExpressionParseException ex) {
+                // This parse error will be reported later. For now, we'll skip this malformed
+                // postcondition and move on to the others.
+                continue;
+            }
+
+            // NOTE: this code is a little inefficient; it creates a single-method annotation and
+            // calls `insertOrRefine` in a loop.  Even worse, this code appears within a loop.  For
+            // now we aren't too worried about it, since the number of
+            // EnsuresCalledMethodsOnException annotations should be small.
+            AnnotationMirror calledMethod =
+                    atypeFactory.createAccumulatorAnnotation(postcond.getMethod());
+            for (Map.Entry<TypeMirror, AccumulationStore> successor :
+                    exceptionalStores.entrySet()) {
+                TypeMirror caughtException = successor.getKey();
+                if (types.isSubtype(caughtException, javaLangExceptionType)) {
+                    AccumulationStore resultStore = successor.getValue();
+                    resultStore.insertOrRefine(e, calledMethod);
+                }
             }
         }
     }
@@ -234,4 +350,15 @@ public class CalledMethodsTransfer extends AccumulationTransfer {
 
         return atypeFactory.createAccumulatorAnnotation(newList);
     }
+
+    /* NO-AFU
+     * Checks if WPI is enabled for the Resource Leak Checker inference. See {@link
+     * ResourceLeakChecker#ENABLE_WPI_FOR_RLC}.
+     *
+     * @return returns true if WPI is enabled for the Resource Leak Checker
+     *
+    protected boolean isWpiEnabledForRLC() {
+        return enableWpiForRlc;
+    }
+    */
 }