+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeValidator.java
@@ -51,6 +51,10 @@ import javax.lang.model.type.TypeVariable;
 /**
  * A visitor to validate the types in a tree.
  *
+ * <p>The validator is called on the type of every expression, such as on the right-hand side of
+ * {@code x = Optional.of(Optional.of("baz"));}. However, note that the type of the right-hand side
+ * is {@code Optional<? extends Object>}, not {@code Optional<Optional<String>>}.
+ *
  * <p>Note: A TypeValidator (this class and its subclasses) cannot tell whether an annotation was
  * written by a programmer or defaulted/inferred/computed by the Checker Framework, because the
  * AnnotatedTypeMirror does not make distinctions about which annotations in an AnnotatedTypeMirror
@@ -298,6 +302,13 @@ public class BaseTypeValidator extends AnnotatedTypeScanner<Void, Tree> implemen
         }
 
         boolean skipChecks = checker.shouldSkipUses(type.getUnderlyingType().asElement());
+        if (type.containsUninferredTypeArguments()) {
+            if (!atypeFactory.ignoreUninferredTypeArguments) {
+                // TODO: document the logic here.
+                isValid = true;
+            }
+            return null;
+        }
 
         if (checkTopLevelDeclaredOrPrimitiveType && !skipChecks) {
             // Ensure that type use is a subtype of the element type