+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -161,7 +161,8 @@ import javax.lang.model.util.ElementFilter;
 
 /**
  * A {@link SourceVisitor} that performs assignment and pseudo-assignment checking, method
- * invocation checking, and assignability checking.
+ * invocation checking, and assignability checking. The visitor visits every construct in a program,
+ * not just types.
  *
  * <p>This implementation uses the {@link AnnotatedTypeFactory} implementation provided by an
  * associated {@link BaseTypeChecker}; its visitor methods will invoke this factory on parts of the
@@ -272,6 +273,9 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     /** True if "-AassumeDeterministic" or "-AassumePure" was passed on the command line. */
     private final boolean assumeDeterministic;
 
+    /** True if "-AassumePureGetters" was passed on the command line. */
+    public final boolean assumePureGetters;
+
     /** True if "-AcheckCastElementType" was passed on the command line. */
     private final boolean checkCastElementType;
 
@@ -344,6 +348,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
                 checker.hasOption("assumeSideEffectFree") || checker.hasOption("assumePure");
         assumeDeterministic =
                 checker.hasOption("assumeDeterministic") || checker.hasOption("assumePure");
+        assumePureGetters = checker.hasOption("assumePureGetters");
         checkCastElementType = checker.hasOption("checkCastElementType");
         conservativeUninferredTypeArguments =
                 checker.hasOption("conservativeUninferredTypeArguments");
@@ -1070,20 +1075,20 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             checkContractsAtMethodDeclaration(
                     tree, methodElement, formalParamNames, abstractMethod);
             /* NO-AFU
-                   // Infer postconditions
-                   if (atypeFactory.getWholeProgramInference() != null) {
-                       assert ElementUtils.isElementFromSourceCode(methodElement);
+            // Infer postconditions
+            if (shouldPerformContractInference()) {
+              assert ElementUtils.isElementFromSourceCode(methodElement);
 
-                       // TODO: Infer conditional postconditions too.
-                       CFAbstractStore<?, ?> store = atypeFactory.getRegularExitStore(tree);
-                       // The store is null if the method has no normal exit, for example if its body is a
-                       // throw statement.
-                       if (store != null) {
-                           atypeFactory
-                                   .getWholeProgramInference()
-                                   .updateContracts(Analysis.BeforeOrAfter.AFTER, methodElement, store);
-                       }
-                   }
+              // TODO: Infer conditional postconditions too.
+              CFAbstractStore<?, ?> store = atypeFactory.getRegularExitStore(tree);
+              // The store is null if the method has no normal exit, for example if its body is a
+              // throw statement.
+              if (store != null) {
+                atypeFactory
+                    .getWholeProgramInference()
+                    .updateContracts(Analysis.BeforeOrAfter.AFTER, methodElement, store);
+              }
+            }
             */
 
             checkForPolymorphicQualifiers(tree.getTypeParameters());
@@ -1093,6 +1098,20 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             methodTree = preMT;
         }
     }
+
+    /* NO-AFU
+     * Should Whole Program Inference attempt to infer contract annotations? Typically, the answer is
+     * "yes" whenever WPI is enabled, but this method exists to allow subclasses to customize that
+     * behavior.
+     *
+     * @return true if contract inference should be performed, false if it should be disabled (even
+     *     when WPI is enabled)
+     */
+    /* NO-AFU
+    protected boolean shouldPerformContractInference() {
+      return atypeFactory.getWholeProgramInference() != null;
+    }
+    */
 
     /**
      * Check method purity if needed. Note that overriding rules are checked as part of {@link
@@ -1131,7 +1150,11 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
         } else {
             r =
                     PurityChecker.checkPurity(
-                            body, atypeFactory, assumeSideEffectFree, assumeDeterministic);
+                            body,
+                            atypeFactory,
+                            assumeSideEffectFree,
+                            assumeDeterministic,
+                            assumePureGetters);
         }
         if (!r.isPure(kinds)) {
             reportPurityErrors(r, tree, kinds);
@@ -1472,7 +1495,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
      * @param qualifier the expression's type, or null if no information is available
      * @return a string representation of the expression and type qualifier
      */
-    private String contractExpressionAndType(
+    protected String contractExpressionAndType(
             String expression, @Nullable AnnotationMirror qualifier) {
         if (qualifier == null) {
             return "no information about " + expression;
@@ -1629,8 +1652,14 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             commonAssignmentCheck(
                     tree, tree.getInitializer(), "enum.declaration.type.incompatible");
         } else if (tree.getInitializer() != null) {
-            // If there's no assignment in this variable declaration, skip it.
-            commonAssignmentCheck(tree, tree.getInitializer(), "assignment.type.incompatible");
+            if (!TreeUtils.isVariableTreeDeclaredUsingVar(tree)) {
+                // If there is no assignment in this variable declaration or it is declared using
+                // `var`, skip it.
+                // For a `var` declaration, TypeFromMemberVisitor#visitVariable already uses the
+                // type of the initializer for the variable type, so it would be redundant to check
+                // for compatibility here.
+                commonAssignmentCheck(tree, tree.getInitializer(), "assignment.type.incompatible");
+            }
         } else {
             // commonAssignmentCheck validates the type of `tree`,
             // so only validate if commonAssignmentCheck wasn't called
@@ -4143,7 +4172,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
     /**
      * Type checks that a method may override another method. Uses an OverrideChecker subclass as
-     * created by {@link #createOverrideChecker}. This version of the method exposes
+     * created by {@link #createOverrideChecker}. This version of the method exposes the
      * AnnotatedExecutableType of the overriding method. Override this version of the method if you
      * need to access that type.
      *
@@ -4184,7 +4213,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     }
 
     /**
-     * Check that a method reference is allowed. Using the OverrideChecker class.
+     * Check that a method reference is allowed. Uses the OverrideChecker class.
      *
      * @param memberReferenceTree the tree for the method reference
      * @return true if the method reference is allowed
@@ -4222,9 +4251,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
         // ========= Overriding Executable =========
         // The ::method element, see JLS 15.13.1 Compile-Time Declaration of a Method Reference
-        ExecutableElement compileTimeDeclaration =
-                (ExecutableElement) TreeUtils.elementFromUse(memberReferenceTree);
-
+        ExecutableElement compileTimeDeclaration = TreeUtils.elementFromUse(memberReferenceTree);
         if (enclosingType.getKind() == TypeKind.DECLARED
                 && ((AnnotatedDeclaredType) enclosingType).isUnderlyingTypeRaw()) {
             if (memRefKind == MemberReferenceKind.UNBOUND) {
@@ -4365,7 +4392,10 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
      */
     public class OverrideChecker {
 
-        /** The declaration of an overriding method. */
+        /**
+         * The declaration of an overriding method. Or, it could be a method reference that is being
+         * passed to a method.
+         */
         protected final Tree overriderTree;
 
         /** True if {@link #overriderTree} is a MEMBER_REFERENCE. */
@@ -4417,10 +4447,10 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             this.overriderTree = overriderTree;
             this.overrider = overrider;
             this.overriderType = overriderType;
+            this.overriderReturnType = overriderReturnType;
             this.overridden = overridden;
             this.overriddenType = overriddenType;
             this.overriddenReturnType = overriddenReturnType;
-            this.overriderReturnType = overriderReturnType;
 
             this.isMethodReference = overriderTree.getKind() == Tree.Kind.MEMBER_REFERENCE;
         }
@@ -4549,7 +4579,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
          */
         protected boolean checkMemberReferenceReceivers() {
             if (overriderType.getKind() == TypeKind.ARRAY) {
-                // Assume the receiver for all method on arrays are @Top
+                // Assume the receiver for all method on arrays are @Top.
                 // This simplifies some logic because an AnnotatedExecutableType for an array method
                 // (ie String[]::clone) has a receiver of "Array." The UNBOUND check would then
                 // have to compare "Array" to "String[]".
@@ -5300,7 +5330,9 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
      */
     protected final boolean shouldSkipUses(ExpressionTree exprTree) {
         // System.out.printf("shouldSkipUses: %s: %s%n", exprTree.getClass(), exprTree);
-
+        // if (atypeFactory.isUnreachable(exprTree)) {
+        //     return true;
+        // }
         Element elm = TreeUtils.elementFromTree(exprTree);
         return checker.shouldSkipUses(elm);
     }