+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractTransfer.java
@@ -83,10 +83,10 @@ import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeMirror;
 
 /**
- * The default analysis transfer function for the Checker Framework propagates information through
- * assignments and uses the {@link AnnotatedTypeFactory} to provide checker-specific logic how to
- * combine types (e.g., what is the type of a string concatenation, given the types of the two
- * operands) and as an abstraction function (e.g., determine the annotations on literals).
+ * The default analysis transfer function for the Checker Framework. It propagates information
+ * through assignments. It uses the {@link AnnotatedTypeFactory} to provide checker-specific logic
+ * to combine types (e.g., what is the type of a string concatenation, given the types of the two
+ * operands) and acts as an abstraction function (e.g., determine the annotations on literals).
  *
  * <p>Design note: CFAbstractTransfer and its subclasses are supposed to act as transfer functions.
  * But, since the AnnotatedTypeFactory already existed and performed checker-independent type
@@ -697,8 +697,14 @@ public abstract class CFAbstractTransfer<
         V elseValue = p.getValueOfSubNode(n.getElseOperand());
         V resultValue = null;
         if (thenValue != null && elseValue != null) {
+            // If a conditional expression is a poly expression, then its Java type is the type of
+            // its context. (For example, the type of the conditional expression in `Object o = b ?
+            // "" : "";` is `Object`, not `String`.)
+            // So, use the Java type of the conditional expression and the annotations for each
+            // branch.
+            TypeMirror conditionalType = TreeUtils.typeOf(n.getTree());
             // The resulting abstract value is the merge of the 'then' and 'else' branch.
-            resultValue = thenValue.leastUpperBound(elseValue);
+            resultValue = thenValue.leastUpperBound(elseValue, conditionalType);
         }
         V finishedValue = finishValue(resultValue, thenStore, elseStore);
         return new ConditionalTransferResult<>(finishedValue, thenStore, elseStore);
@@ -1017,7 +1023,7 @@ public abstract class CFAbstractTransfer<
     @Override
     public TransferResult<V, S> visitDeconstructorPattern(
             DeconstructorPatternNode n, TransferInput<V, S> in) {
-        // TODO: dummy implementation to avoid crash.
+        // TODO: Implement getting the type of a DeconstructorPatternTree.
         V value = null;
         return createTransferResult(value, in);
     }
@@ -1061,7 +1067,7 @@ public abstract class CFAbstractTransfer<
      * @return whether to perform whole-program inference on the tree
      */
     /* NO-AFU
-    private boolean shouldPerformWholeProgramInference(Tree tree) {
+    protected boolean shouldPerformWholeProgramInference(Tree tree) {
       TreePath path = this.analysis.atypeFactory.getPath(tree);
       return infer && (tree == null || !analysis.checker.shouldSuppressWarnings(path, ""));
     }
@@ -1076,7 +1082,7 @@ public abstract class CFAbstractTransfer<
      * @return whether to perform whole-program inference
      */
     /* NO-AFU
-    private boolean shouldPerformWholeProgramInference(Tree expressionTree, Tree lhsTree) {
+    protected boolean shouldPerformWholeProgramInference(Tree expressionTree, Tree lhsTree) {
       // Check that infer is true and the tree isn't in scope of a @SuppressWarnings
       // before calling InternalUtils.symbol(lhs).
       if (!shouldPerformWholeProgramInference(expressionTree)) {